Move table of contents to left hand side
The answer is in `layouts/_default/list.html` (line 17) and `layouts/_default/single.html` (line 38) files where we need to remove `lg:order-last` in the class of the HTML element. I had to adjust those files and add them to our project. In addition, I had to adjust the CSS and added .ps-5 selector in custom.css